+++ tests/integration/test_magicc7.py
@@ -1,5 +1,6 @@
 import numpy.testing as npt
 import pymagicc.io
+import pytest
 from scmdata import ScmRun
 
 from openscm_runner import run
@@ -18,7 +19,10 @@ def test_magicc7_run(test_scenarios, magicc7_is_available):
                     "rf_soxi_dir_wm2": -0.2,
                     "out_temperature": 1,
                     "out_forcing": 1,
-                    "out_dynamic_vars": ["DAT_AEROSOL_ERF"],
+                    "out_dynamic_vars": [
+                        "DAT_AEROSOL_ERF",
+                        "DAT_HEATCONTENT_AGGREG_TOTAL",
+                    ],
                     "out_ascii_binary": "BINARY",
                     "out_binary_format": 2,
                 },
@@ -46,9 +50,9 @@ def test_magicc7_run(test_scenarios, magicc7_is_available):
             "Effective Radiative Forcing",
             "Effective Radiative Forcing|Aerosols",
             "Effective Radiative Forcing|CO2",
+            "Heat Content|Ocean",
             # "CO2 Air to Land Flux",  # todo: add this back in
         ),
-        full_config=False,
     )
 
     assert isinstance(res, ScmRun)
@@ -63,7 +67,21 @@ def test_magicc7_run(test_scenarios, magicc7_is_available):
             "Effective Radiative Forcing",
             "Effective Radiative Forcing|Aerosols",
             "Effective Radiative Forcing|CO2",
+            "Heat Content|Ocean",
         ]
+    )
+
+    # check ocean heat content unit conversion comes through correctly
+    npt.assert_allclose(
+        1844.7,
+        res.filter(
+            unit="ZJ",
+            variable="Heat Content|Ocean",
+            region="World",
+            year=2100,
+            scenario="ssp126",
+        ).values.max(),
+        rtol=RTOL,
     )
 
     npt.assert_allclose(
@@ -183,3 +201,52 @@ def test_write_scen_files_and_make_full_cfgs(
             scen_flag_val = scenario_cfg["file_emisscen_{}".format(i)]
 
             assert scen_flag_val == "NONE"
+
+
+@pytest.mark.parametrize(
+    "out_config",
+    (
+        ("core_climatesensitivity", "rf_total_runmodus"),
+        ("core_climatesensitivity",),
+        ("rf_total_runmodus",),
+    ),
+)
+def test_return_config(test_scenarios, magicc7_is_available, out_config):
+    core_climatesensitivities = [2, 3]
+    rf_total_runmoduses = ["ALL", "CO2"]
+
+    cfgs = []
+    for cs in core_climatesensitivities:
+        for runmodus in rf_total_runmoduses:
+            cfgs.append(
+                {
+                    "out_dynamic_vars": [
+                        "DAT_TOTAL_INCLVOLCANIC_ERF",
+                        "DAT_SURFACE_TEMP",
+                    ],
+                    "core_climatesensitivity": cs,
+                    "rf_total_runmodus": runmodus,
+                }
+            )
+
+    res = run(
+        climate_models_cfgs={"MAGICC7": cfgs},
+        scenarios=test_scenarios.filter(scenario=["ssp126", "ssp245", "ssp370"]),
+        output_variables=("Surface Temperature", "Effective Radiative Forcing",),
+        out_config={"MAGICC7": out_config},
+    )
+
+    for k in out_config:
+        assert k in res.meta.columns
+        ssp126 = res.filter(scenario="ssp126")
+
+        # check all the configs were used and check that each scenario
+        # has all the configs included in the metadata too
+        if k == "core_climatesensitivity":
+            assert set(res.get_unique_meta(k)) == set(core_climatesensitivities)
+            assert set(ssp126.get_unique_meta(k)) == set(core_climatesensitivities)
+        elif k == "rf_total_runmodus":
+            assert set(res.get_unique_meta(k)) == set(rf_total_runmoduses)
+            assert set(ssp126.get_unique_meta(k)) == set(rf_total_runmoduses)
+        else:
+            raise NotImplementedError(k)